requirements.txt
flask
psycopg2-binary
nanoid
python-dotenv
gunicorn  